+++ app/src/main/AndroidManifest.xml
@@ -9,14 +9,14 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".features.onboarding.OnboardingActivity">
+        <activity android:name=".ui.onboarding.OnboardingActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
 
-        <activity android:name=".features.group.GroupActivity">
+        <activity android:name=".ui.group.GroupActivity">
             <intent-filter>
                 <action android:name="android.intent.action.VIEW" />
