+++ app/src/main/java/com/ksss/splintter/data/business/GroupBo.java
@@ -1,0 +1,14 @@
+package com.ksss.splintter.data.business;
+
+import android.support.annotation.NonNull;
+import com.ksss.splintter.data.business.exception.EmptyNameException;
+import com.ksss.splintter.data.business.exception.NameTooShortException;
+import com.ksss.splintter.data.model.Group;
+
+/**
+ * Created by Nahuel Barrios on 7/19/16.
+ */
+public interface GroupBo {
+
+    Group create(@NonNull String name) throws EmptyNameException, NameTooShortException;
+}
