+++ app/src/main/java/com/ksss/splintter/data/business/exception/NameTooShortException.java
@@ -1,4 +1,4 @@
-package com.ksss.splintter.features.group.backend.exception;
+package com.ksss.splintter.data.business.exception;
 
 /**
  * Created by Nahuel Barrios on 7/19/16.
