--- app/src/main/java/com/ksss/splintter/features/group/backend/PersonBO.java
@@ -1,15 +1,0 @@
-package com.ksss.splintter.features.group.backend;
-
-import android.support.annotation.NonNull;
-import com.ksss.splintter.features.group.backend.exception.EmptyNameException;
-import com.ksss.splintter.features.group.backend.exception.NameTooShortException;
-import com.ksss.splintter.features.group.domain.Person;
-
-/**
- * Created by Nahuel Barrios on 7/19/16.
- */
-public interface PersonBo {
-
-    @NonNull
-    Person create(@NonNull String name) throws EmptyNameException, NameTooShortException;
-}
