+++ app/src/main/java/com/ksss/splintter/ui/group/view/GroupExpensesFragment.java
@@ -1,4 +1,4 @@
-package com.ksss.splintter.features.group.view;
+package com.ksss.splintter.ui.group.view;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
@@ -17,15 +17,15 @@ import android.widget.TextView;
 import com.joanzapata.iconify.IconDrawable;
 import com.joanzapata.iconify.fonts.MaterialIcons;
 import com.ksss.splintter.R;
-import com.ksss.splintter.features.group.backend.ExpenseBo;
-import com.ksss.splintter.features.group.backend.PersonBo;
-import com.ksss.splintter.features.group.backend.exception.EmptyNameException;
-import com.ksss.splintter.features.group.backend.exception.NameTooShortException;
-import com.ksss.splintter.features.group.backend.impl.ExpenseBoImpl;
-import com.ksss.splintter.features.group.backend.impl.PersonBoImpl;
-import com.ksss.splintter.features.group.domain.Expense;
-import com.ksss.splintter.features.group.domain.Group;
-import com.ksss.splintter.features.group.domain.Person;
+import com.ksss.splintter.data.business.ExpenseBo;
+import com.ksss.splintter.data.business.PersonBo;
+import com.ksss.splintter.data.business.exception.EmptyNameException;
+import com.ksss.splintter.data.business.exception.NameTooShortException;
+import com.ksss.splintter.data.business.impl.ExpenseBoImpl;
+import com.ksss.splintter.data.business.impl.PersonBoImpl;
+import com.ksss.splintter.data.model.Expense;
+import com.ksss.splintter.data.model.Group;
+import com.ksss.splintter.data.model.Person;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import hugo.weaving.DebugLog;
 import io.realm.Realm;
