+++ app/src/main/java/com/ksss/splintter/ui/group/view/GroupExpensesSummaryFragment.java
@@ -1,4 +1,4 @@
-package com.ksss.splintter.features.group.view;
+package com.ksss.splintter.ui.group.view;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -6,7 +6,6 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
 import com.ksss.splintter.R;
 
 /**
